cloud_reid_functions/event_handler/requirements.txt
flask
google-cloud-pubsub
google-cloud-datastore
pydantic
